Fix webhint error
