+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-  mode: 'production',
+  mode: 'development',
   entry: {
     bundle: path.resolve(__dirname, 'src/index.js'),
   },
